+++ client/webpack/package.json
@@ -19,6 +19,7 @@
     "html-loader": "^0.5.5",
     "moment": "^2.24.0",
     "vue": "^2.5.2",
+    "vue-file-agent": "^1.7.3",
     "vue-friendly-iframe": "^0.16.0",
     "vue-json-excel": "^0.2.98",
     "vue-material-design-icons": "^3.3.1",
