+++ client/webpack/src/store/model/sliders.js
@@ -35,7 +35,6 @@ const module = {
       curl.get('/slider_img')
         .then(function(response) {
           // handle success
-          console.log("Obtenido de AWS:",response.data)
           if(response.data){
             commit('updateSliderHomeItems',response.data)
           }else{
