+++ server/index.js
@@ -59,6 +59,8 @@ const generalRoutes = require('./routes/generals');
 const descargasRoutes = require('./routes/descargas');
 const matriculaRoutes = require('./routes/matricula');
 const sliderRoutes = require('./routes/slider_img');
+const uploaderRoutes = require('./routes/uploader');
+const mailerRoutes = require('./routes/mailer');
 
 // LAS RUTAS
 app.use('/users', usersRoutes);
@@ -75,6 +77,8 @@ app.use('/general',generalRoutes);
 app.use('/descargas',descargasRoutes);
 app.use('/matricula',matriculaRoutes);
 app.use('/slider_img',sliderRoutes);
+app.use('/upload',uploaderRoutes);
+app.use('/mailer',mailerRoutes);
 
 app.get('/', (request, response) => {
   response.json({ info: 'VueJs, Node.js, Express, and WordPress-API' })
