+++ server/package.json
@@ -21,16 +21,25 @@
     "cors": "^2.8.5",
     "dotenv": "^8.0.0",
     "express": "^4.16.4",
+    "express-fileupload": "^1.1.7-alpha.3",
     "jsonwebtoken": "^8.5.1",
     "mercadopago": "^1.1.0",
+    "morgan": "^1.10.0",
+    "multer": "^1.4.2",
+    "multer-ftp": "^1.2.0",
+    "multer-sftp": "^1.1.0",
     "mysql": "^2.17.1",
     "mysql2": "^1.6.5",
     "node-localstorage": "^1.3.1",
+    "nodemailer": "^6.4.8",
     "nodemon": "^1.19.0",
+    "path": "^0.12.7",
     "pg": "^7.10.0",
     "pg-hstore": "^2.3.2",
+    "qrcode": "^1.4.4",
     "request": "^2.88.0",
-    "sequelize": "^5.8.5",
+    "sequelize": "^5.21.11",
+    "ssh2-sftp-client": "^5.1.2",
     "wpapi": "^1.2.1"
   }
 }
